feat: fetch 응답 본문 없을 때 response 바로 반환하기, 일정 삭제 API 연결
