+++ frontend/src/features/my-calendar/api/index.ts
@@ -21,4 +21,7 @@ export const personalEventApi = {
     const response = await request.put(`/api/v1/personal-event/${id}`, { body });
     return response;
   },
+  deletePersonalEvent: async (id: number): Promise<void> => {
+    await request.delete(`/api/v1/personal-event/${id}`);
+  },
 };
