+++ frontend/src/utils/fetch/index.ts
@@ -54,8 +54,12 @@ export const executeFetch = async (
       throw new Error(`HTTP Error: ${response.status}`);
     }
 
+    const text = await response.clone().text();
+    if (!text) return response;
+
     const data = await response.json();
     return data;
+
   } catch (error) {
     throw new Error(`Network Error : ${error}`);
   }
